style: fix chat button, margin, userItem margin
