+++ src/api/chatBotApi.js
@@ -33,4 +33,19 @@ const postChat = async (question, token) => {
   return res.json();
 }
 
-export { getChat, postChat };
+const getStudentChat = async (studentId, token) => {
+  const res = await fetch(import.meta.env.VITE_APP_CHAT_API_URL + '/api/chat/student?studentId=' + studentId, {
+    headers: {
+      'authorization': 'Bearer ' + token
+    }
+  });
+
+  if (!res.ok) {
+    const message = (await res.json()).message;
+    throw new Error(message);
+  }
+
+  return res.json();
+}
+
+export { getChat, postChat, getStudentChat };
