+++ src/routes/ChatStu.css
@@ -141,16 +141,3 @@
 .input-wrapper .input-text::-webkit-scrollbar {
     display: none;
 }
-
-.send-btn {
-  border: none;
-  background-color: #001832;
-  border-radius: 50%;
-  cursor: pointer;
-  margin: 5px;
-}
-
-.send-btn img {
-  width: 35px; 
-  height: 35px; 
-}
