+++ src/routes/History.css
@@ -45,6 +45,7 @@
   margin: 56px auto;
   max-width: 800px;
   width: 80%;
+  margin-bottom: 120px;
 }
 
 .message-wrapper {
